Fixed compilation.
Everything now compiles, but I have tested only the rand- stuff.
Also, the log-file might not be working right.
Executables have been renamed to main.go to make it obvious where
the `main` func is, and to aid tab-completion in the dirs.
