+++ examples/GetWidgets/initiators/web/Makefile
@@ -5,7 +5,7 @@ include $(GOROOT)/src/Make.inc
 
 TARG=web
 GOFILES=\
-	web.go\
+	main.go\
 
 # gb: this is the local install
 GBROOT=../..
